rename from _ to - name style.  rename -schema to .schema to more closely link with the associated .data file.
