+++ src/lib/map-string-hex.js
