+++ src/lib/unit/unit-attrs.js
@@ -2,7 +2,7 @@ const assert = require('../../wrapper/assert')
 const _ = require('../../wrapper/lodash')
 const locale = require('../locale')
 const { ObjectNamespace } = require('../object-namespace')
-const { UnitAttrsSchema } = require('./unit-attrs-schema')
+const { UnitAttrsSchema } = require('./unit-attrs.schema')
 const { UnitModifier } = require('./unit-modifier')
 const UNIT_ATTRS = require('./unit-attrs.data')
 const { world, Card, GameObject } = require('../../wrapper/api')
